+++ addon/util/socket-client-lookup.js
@@ -1,3 +1,0 @@
-export default function(owner, socketClientName) {
-  return owner.lookup(`ember-socket-guru@socket-client:${socketClientName}`);
-}