+++ tests/dummy/app/services/socket-guru-action-cable.js
@@ -1,9 +1,0 @@
-import SocketGuru from 'ember-socket-guru/services/socket-guru';
-
-export default SocketGuru.extend({
-  socketClient: 'action-cable',
-  config: {
-    socketAddress: 'ws://0.0.0.0:28080',
-  },
-  observedChannels: ['CommentsChannel'],
-});